+++ CHANGELOG.md
@@ -121,4 +121,5 @@
 - Made it so that removing Overclockers from the Advanced Incubator will reset its progress
 - Fixed the Incubator not being able to output
 - Fixed the Plasmid Infuser giving +1 DNA Point for Helices of the same Gene rather than +2 (#18)
+- Fixed the Advanced Incubator ignoring Chorus Fruit (#17)
 - In the Plasmid Infuser, the initial DNA Helix infused into an empty Plasmid does not give it any DNA Points, previously it gave +1